infra/lambda-function/lambda-proxy-oc/requirements.txt
flask
mangum
psycopg2-binary
cuid2